backend/pyproject.toml
[project]
name = "pool-temperature"
version = "0.1.0"
description = "Swimming Pool Temperature"
requires-python = ">=3.13"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "mysql-connector-python",
    "requests",
]